% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_main.R
\name{BarycentricInterpolation.main}
\alias{BarycentricInterpolation.main}
\title{TODO:
Add a function for BarycentricInterpolation(xvals, yvals, weightfunc)
Add a function for BarycentricInterpolation.Polynomial(xvals,yvals)
Add a function for BarycentricInterpolation.Chebyshev(xvals,yvals)
Use the new barycentric interpolation functions to simplify previous code for approximations
Debug the barycentric interpolation function
Add a function for FAST interpolation with BarycentricInterpolation.Berrut(xvals,yvals)}
\usage{
BarycentricInterpolation.main(xvals, yvals, weights)
}
\arguments{
\item{weights}{The weights to interpolate on}

\item{xvec}{A vector of the x values}

\item{yvec}{A vector of the y values}
}
\description{
Bayes Stuff:
Add a function that computes, given only value of x's as input:
 gamma posterior for 
 lambda posterior for number of arrivals
 mean & var posterior
 var posterior for population
 pareto posterior for uniform
 dirlecht distribution for categorical variance
 beta posterior for probability
}
\details{
Compute the Barycentric Interpolant for a given weight function

Take in a vector of x,y pairs and a weight function, and return a function that provides the barycentric interpolant through those points. This function is guaranteed to be rational
}
\examples{
 
f <- function(x) {
  sin(x)
}

x <- seq(0,1,length = 100)
# Caution, this interpolatnt is highly unstable!!!
curve(barycentricInterpolation(x, f(x), function(xvals, yvals) {return(-1)^length(x)})
}
\keyword{interpolate}
\keyword{interpolation}
\keyword{rational}

