% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_main.R
\name{chebfun}
\alias{chebfun}
\title{Approximate a function with a Chebyshev Series}
\usage{
chebfun(fxn, lower, upper, tol = 1e-10, maxN = 10000, verbose = F)
}
\arguments{
\item{fxn}{The function we wish to approximate}

\item{lower}{The lower bound of the approximation of the range}

\item{upper}{The upper bound of the approximation of the range}

\item{tol}{Any coefficient smaller than this will be disreguarded}

\item{maxN}{The maximum number of coefficients to calculate before the 
program errors and says there is no resolution to this approximation}

\item{verbose}{logical; if true, the chebyshev_coef function will print a 
verbose message output indicating the coefficient sizes and the amount of 
digits (base 10) that each coefficient represents}
}
\description{
This function takes in a minimum and a maximum range, a function, and a 
tolerance, and returns a chebyshev series evalued based on the coefficients
of the series. This function is actually faster than the interpolation
function for a simular value of N
Sources: SIAM Chebyshev Expansions: EQ3.62
for coefficients Algorithm 3.2 for evaluating the sum (Modified Clenshaws
algorithm) EQ3.70 for error bounds
}
\examples{
f <- function(x) {
   x^5
}
chebfun(f, N = 1, lower = l, upper = u, tol = 1E-15)
}
\keyword{approximation}
\keyword{chebyshev}

