% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_main.R
\name{chebyshev_bestN}
\alias{chebyshev_bestN}
\title{Compute the the value of N for which a good chebyshev approximation would be}
\usage{
chebyshev_bestN(fxn, lower, upper, tol = 1e-10, maxN = 10000)
}
\arguments{
\item{fxn}{The function we wish to approximate}

\item{lower}{The lower bound of the approximation of the range}

\item{upper}{The upper bound of the approximation of the range}

\item{tol}{Any coefficient smaller than this will be disreguarded}

\item{maxN}{The maximum number of coefficients to calculate before the
program errors and says there is no resolution to this approximation}
}
\description{
This function takes in a minimum and a maximum range, a function, and a
tolerance, and returns a the chebyshev coefficients of a function - useful
for determining the proper N to use. Sources: SIAM Chebyshev Expansions EQ3.62
#' @keywords chebyshev approximation
}
\examples{
f <- function(x) {
   x^5
}
chebyshev_bestN(f, N = 1, lower = l, upper = u, tol = 1E-15)
Use this code which does the following:
1) Find out which of the coefficients are less than the tolerance
2) Find which elements of 1:length(coefs) are not less than the tolerance
3) Use the values from (2) to determine the index of the coefficient that
   is not
}

