% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_main.R
\name{chebyshev_coef}
\alias{chebyshev_coef}
\title{Compute the Chebyshev Coefficients for a given function}
\usage{
chebyshev_coef(fxn, lower, upper, tol = 1e-10, maxN = 10000, verbose = F)
}
\arguments{
\item{fxn}{The function we wish to approximate}

\item{lower}{The lower bound of the approximation of the range}

\item{upper}{The upper bound of the approximation of the range}

\item{tol}{Any coefficient smaller than this will be disreguarded}

\item{maxN}{The maximum number of coefficients to calculate before the
program errors and says there is no resolution to this approximation}

\item{verbose}{logical; if true, the chebyshev_coef function will print a
verbose message output indicating the coefficient sizes and the amount of
digits (base 10) that each coefficient represents}
}
\description{
This function takes in a minimum and a maximum range, a function, and a
tolerance, and returns a the chebyshev coefficients of a function - useful
for determining the proper N to use. Sources: SIAM Chebyshev Expansions
EQ3.62 #' @keywords chebyshev approximation
}
\examples{
f <- function(x) {
   x^5
}
chebyshev_coef(f, lower = -2, upper = 2, tol = 1E-10)
}

