% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_main.R
\name{chebyshev_compute}
\alias{chebyshev_compute}
\title{Chebyshev Interpolation Function}
\usage{
chebyshev_compute(fxn, lower, upper, tol = 1e-10, n_subdivisions = 100,
  maxN = 3000, maxEfficiency = F)
}
\arguments{
\item{fxn}{The function we wish to approximate}

\item{lower}{The lower bound of the approximation of the range}

\item{upper}{The upper bound of the approximation of the range}

\item{tol}{Tolerance of the functional approximation}
}
\description{
This function takes in a minimum and a maximum range, a function, and a
tolerance, and returns a chebyshev approximation to a function
}
\examples{
f <- function(x) {
   x^5
}
chebyshev_compute(f, lower = -2, upper = 2, tol = 1E-10)
}
\keyword{approximation}
\keyword{chebyshev}

