% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functtions.R
\name{gibbs_sampler}
\alias{gibbs_sampler}
\title{Gibbs Sampling for Bivariate Normal Distribution}
\usage{
gibbs_sampler(mu, sigma, start, n_samples)
}
\arguments{
\item{mu}{Vector of means for the bivariate normal distribution.}

\item{sigma}{Covariance matrix for the bivariate normal distribution.}

\item{start}{Initial values for the chain.}

\item{n_samples}{Number of samples to generate.}
}
\value{
A matrix of samples.
}
\description{
Gibbs Sampling for Bivariate Normal Distribution
}
\examples{
mu <- c(0, 0)
sigma <- matrix(c(1, 0.5, 0.5, 1), nrow = 2)
samples <- gibbs_sampler(mu, sigma, start = c(0, 0), n_samples = 1000)
}
