% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_main.R
\name{meroSFunc}
\alias{meroSFunc}
\title{Meromorphic function for function weighting}
\usage{
meroSFunc(x)
}
\arguments{
\item{x}{The input value to the function x \in [0,1] '}
}
\description{
Meromorphic Sinosoid function that makes this possible Very cool properties - bounded, parameterized, monotonic, C^Inf ' Sources: Ulrich Multz Documentation on using this function for a parabolic interpolation program
}
\examples{
 
meroSFunc(0:100 / 100)
}
\keyword{meromorphic}
\keyword{parabola}
\keyword{weight}
\keyword{weighting}

