% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functtions.R
\name{metropolis_hastings}
\alias{metropolis_hastings}
\title{Metropolis-Hastings Sampling}
\usage{
metropolis_hastings(target_density, proposal_sd, start, n_samples)
}
\arguments{
\item{target_density}{Function for the target density to sample from.}

\item{proposal_sd}{Standard deviation of the proposal distribution.}

\item{start}{Initial value for the chain.}

\item{n_samples}{Number of samples to generate.}
}
\value{
A numeric vector of samples.
}
\description{
Metropolis-Hastings Sampling
}
\examples{
target <- function(x) { dnorm(x, mean = 0, sd = 1) }
samples <- metropolis_hastings(target, proposal_sd = 1, start = 0, n_samples = 1000)
}
